better view of event logs
